Disable implicit extensibility warning for tests
In most tests the extensibility of the types used is not relevant.
Therefore, the IDL files do not explicitly set the extensibility and
use the default. IDLC warns about implicit extensibility because the
default (currently final) may change to appendable (in the spec).
For all cases where this warning is not relevant (because a specific
extensibility is not needed), this commit disables the warning.

Signed-off-by: Dennis Potman <dennis.potman@adlinktech.com>