Fix get_typelookup_writer racing new_participant
A new (local) participant is made visible in the entity index before
creating the built-in writers, to avoid complications in the matching.
This is generally fine, they don't get used until later on anyway,
except in those cases where one tries to use a built-in writer by simply
scanning all the participants.  And this is what the
get_typelookup_writer does.

The result is that calling get_builtin_writer may find that the
requested writer doesn't exist (yet, in this case), causing it to fall
back to the "privileged participant" without that "privileged
participant" being set for the participant.  This causes an assertion
failure.

This commit replaces a boolean value tracking the progress of deleting
the participant to a enum that allows one to mark also check progress in
initialization.  Any participant not yet ready is then simply skipped,
avoiding the race.

Signed-off-by: Erik Boasson <eb@ilities.com>