Bounded sequences
Adds the implementation of bounded sequences in IDLC and the
cdrstream serializer. A new serializer instruction BSQ is
introduced, that has a bound parameter and is handled similar
to the existing SEQ instruction.

Serialization of a sample that has a sequence-type member containing
more elements than the specified bound will fail. For incoming
data, deserialization (normalize) also fails in case a sequence
has an elements count that is larger than its bound (the sample
will be dropped in that case).

Signed-off-by: Dennis Potman <dennis.potman@adlinktech.com>