Fix data representation in RTPS encapsulation identifier
The RTPS encapsulation identifier should correspond to the encoding
version and data representation of the outermost object of the type
used for a topic. This commit fixes the issue that Cyclone was using
an incorrect value for this identifier when writing data for a topic
with an appendable or mutable top-level type.

The Cyclone deserializer uses the data representation as provided
in the serializer instructions for that type, but did not check that
the encapsulation identifier in the RTPS submessage indicates the
same data representation. This commit also introduces that check.

Signed-off-by: Dennis Potman <dennis.potman@adlinktech.com>