Do not force XCDR1 default for DATA_REPRESENTATION
Simply merging in the default reader/writer QoS causes it to pick up the
hard-coded XCDR1 that is specified for the discovery protocol.  This
then means that "ensure_valid_data_representation" can only accept or
reject the setting, instead of being able to pick a sensible default
based on the type definition.  The end result is that an application
would be forced to explicitly select XCDR2 for topics that require it.
That's not good.

Signed-off-by: Erik Boasson <eb@ilities.com>