Rename config_init, &c to ddsi_config_init, &c
The old names cause name clashes with libconfig (cf GitHub issue #1133)
which, combined with a general sense that it would be better to prefix
globally visible names, renaming these functions to have a ddsi_ prefix
seems the most sensible.

It furthermore renames the q_config files because cleaning up symbols
names and ending up with a file name with a different prefix is not a
good plan.

Signed-off-by: Erik Boasson <eb@ilities.com>