Use liveliness QoS for participant lease duration
The DDSI spec defines a "participant lease duration" at the protocol
level without defining a mechanism for configuring it.  Traditionally
Cyclone DDS made it possible to configure it via the LeaseDuration
setting, this commit adds the possibility for an application to
manipulate it via the participant QoS.

Given that the spec says the liveliness QoS is applicable only to the
topic, reader and writer, and that the "participant lease duration"
exactly matches what one expects from an "automatic" liveliness kind, it
seems perfectly reasonable to configure this via the liveliness QoS on
the participant.  This liveliness setting is then also propagated into
the built-in topics.

This does result in a change in behaviour for those applications that
create a single QoS object containing a liveliness setting that they
then use for creating the participant and a reader/writer.  Without
these changes, the participant would ignore it, now it will use the
value (or reject it, if it is not "automatic").

Signed-off-by: Erik Boasson <eb@ilities.com>