Implement @data_representation in DDSI
A topic descriptor can have a bitmask of allowed data representations for a type, which is
included in case the @data_representation annotation is used on the top-level type. This commit
adds the implementation for this restriction in DDSI. A test case is included to test the
combination of topic qos and data_representation annotation.

Signed-off-by: Dennis Potman <dennis@zettascale.tech>