Set CMake CMP0075 policy to NEW
CMP0075 concerns a change in behaviour in the combination of
check_include_files and CMAKE_REQUIRED_LIBRARIES. This combination
occurs on Windows, and in this case both the old and the new behaviour
are acceptable.

We support CMake versions older than the one where the policy was
introduced, so select the new behaviour only if the policy actually
exists.

Signed-off-by: Erik Boasson <eb@ilities.com>