Add options for output directory and path reconstruction (#1220)
Added -o command line option to specify an output directory which may or may not exist.
Added -b command line option to specify a base path for a directory structure.
When using -b, any directories from <base_dir> to <file> will be reconstructed in the output directory. If no output directory is specified, this will result in files being compiled in-place.
Necessary directories will be constructed automatically to hold the output files

Signed-off-by: Andrew Hoener <ahoener@boston-engineering.com>