Use only functions that are aware of loaned samples when dealing with loaned pointers
Loaned samples are just pointers to the iceoryx buffers. CDDS allows
bindings to wrap the sample in their own structure.  A binding needs
to know when the sample pointer is a loaned pointer so it doesn't
attempt to dereference it as a custom sample structure.

When iceorx is available on a reader we route the calls through the
ddsi_serdata_from_loaned_sample. The binding will know to handle
it properly.