ADLINK Technology Ltd became ZettaScale Technology
This updates the copyright comments at the beginning of the files to
reflect this change.

It turns out in many cases the years weren't updated when editing, so
I've tried to get the correct years in there based on the git commit
log.

There are some files where one line is removed, in those files two lines
(A and Z) are replaced those by a single one (Z).

Signed-off-by: Erik Boasson <eb@ilities.com>