Remove Discovery/Peers/Group and all related code
This setting was inherited from OpenSplice, where it was useful in
combination with another product that hasn't been available for a number
of years.

Deprecating the option would technically be the correct route, but in
this situation, it seems like no-one can ever have actually used it and
therefore deleting it outright is reasonable.

Signed-off-by: Erik Boasson <eb@ilities.com>