Use C11 _Alignof function for alignment
All supported compilers have support for the C11 _Alignof keyword,
so we can rely on this for the alignment used in the topic descriptor.
This replaces the custom implementation of alignment calculation
in idlc, that in certain cases reported the wrong alignment when
using external or optional members.

This also replaces the alignof macro that was defined and used in
whc and plist code with the dds_alignof macro added in ddsrt.

Signed-off-by: Dennis Potman <dennis@zettascale.tech>