Update "perftest" script configuration settings
This script was overlooked when updating the configuration settings on
two points:

* Moving the default lease duration to the Discovery section
* Improving the way socket buffer sizes get specified

The old settings are still handled, but it shouldn't trigger any
deprecation warnings.  The new way of specifying socket receive buffer
sizes also distinguishes between what to ask for and what the require.
This should prevent an error on many Linux systems in the default
configuration.

Signed-off-by: Erik Boasson <eb@ilities.com>