Fix by source timestamp with more than 2 writers
Thanks to @secondage pointing out a bit of dead code (and some luck in
it ultimately getting the attention it really deserved), this commit can
now squash a bug in the "by source timestamp" destination order QoS
setting implementation.

By not setting the writer GUID associated with the latest sample pushed
into the writer, the logic checking whether the GUID of the writer or an
arriving sample with the same timestamp as the most recently insert one
can't do its job.

This adds some test cases, including one that showed the issue. They are
merely a start for writing some more tests where the effects of
various (combinations of) QoS settings get tested in an otherwise
reasonably isolated manner.

Signed-off-by: Erik Boasson <eb@ilities.com>