Fix issue in IDLC related to nested include files
This fixes the issue in IDLC that when using input files that have nested
include files, the conversion of the relative to absolute path in IDLC is
incorrect when popping sources files from the stack in the push_line
function. This issue resulted in no output from IDLC, but it also didn't
report an error.

This commit changes the way files are popped from the stack based on
the processor filemakers: it assumes that for every START_OF_FILE there
is a corresponding RETURN_TO_FILE to return to the parent file, and
therefore there is no need to convert the relative path to an absolute
path when processing a RETURN_TO_FILE marker.

Signed-off-by: Dennis Potman <dennis@zettascale.tech>