Do not lookup_thread_state before pinning entity
Calling lookup_thread_state results in a crash if the library hasn't
been initialized yet, which happens as a consequence of instantiating a
domain.

All operations that need to lookup the thread state are in some way
related to an entity, and pinning an entity does check for an
uninitialized library.

This commit simply moves the handful of calls to lookup_thread_state.