Avoid minimal type ids in ddsi_type_get_typeinfo
The requirement on input is a complete type id that refers to a fully
resolved type, and then it constructs both the complete and the minimal
type information using that complete type.

The problem with the implementation was that the generator was
parameterized for the output kind, and called once for minimal and once
for complete type information.  Constructing the dependent type
information can require recursion, in which case it would recurse using
the just-generated minimal type ids.  Those ids need not exist yet.

This commit addresses that problem by generating both in one go, always
recursing using the complete type id.

As the number of dependent hashed type ids may be different for the
complete and the mininal type information, it now also de-dups the
dependent type ids.

Signed-off-by: Erik Boasson <eb@ilities.com>