Make object to add to hash table non-const
GCC's static analyzer assumes no ownership of allocated is taken when
the object is passed to a const pointer, see for example

  https://gcc.gnu.org/bugzilla/show_bug.cgi?id=101713

describing the exact same case where the object is merely stored in the
hash table, and the owner of the hash table remains responsible for
freeing the memory.

Here is arguably is better to make the argument non-const anyway,
because the only way of getting the object out of the hash table
involves casting away the const.

Note that the "ehh" case is different, it copies the data.

It also turns out that changing this doesn't affect the rest of the code
base other than some wrapper functions in test code.

Signed-off-by: Erik Boasson <eb@ilities.com>