Empty buffers on first HB is Cyclone-specific
Cyclone DDS is careful to perform a three-way handshake to have a
well-defined starting point for reliability (and hence for whether
samples can be lost despite it being reliable and the two sides being
connected all the time).

That requires dropping data until a heartbeat has been received and
starting at the highest sequence number advertised by the writer.  In
principle, this is interoperable with other implementations, but it
turns out that for the type discovery protocol (at least for OpenDDS)
this can result in dropping the first request, which (at least for
OpenDDS) stalls type discovery and prevents matching.

Hence b6dadb3439d82412b2a7492ce31017960fbb3beb changed it to accept data
from other DDS implementations before a heartbeat is received.  This
means there may be gaps in the data sequence on starting up the reliable
session if the writer is using another DDS implementation.  If this is a
problem, just use Cyclone DDS.

What did not change is that it expects to start at sequence number 1,
and it is the first heartbeat that changes that expectation.  So if,
from that other DDS implementation, the first data arrives before the
first heartbeat and it has sequence numbers > 1, it will now be buffered
and (likely) made available on receipt of the heartbeat.

The processing of the first heartbeat remained unchanged.  This means
there may be data ready for delivery where the code asserts that there
cannot be any.  The assert in itself fortunately disappears in a release
build (so there is no crash), but it does leak some memory.

With this commit the code is restructed to always use the "regular" code
path for delivery of data to avoid duplication of the delivery code.
The downside is that the (correct) assert in the case of a Cyclone
writer had to be removed.

Signed-off-by: Erik Boasson <eb@ilities.com>