macOS: if Iceoryx used, add library dir to rpath
The CMAKE_INSTALL_RPATH set in the top-level CMakeLists.txt causes the
rpath for the Iceoryx library to disappear on installation, but not
setting it means the build directories are used in the rpaths of the
installed Cyclone library, which also doesn't work.

This adds the directory where the iceoryx library is found to the rpath
used for installing.  I suspect there are better ways, but this should
be better than nothing.

Signed-off-by: Erik Boasson <eb@ilities.com>