Increase time-out for topic discovery test
Under high load, with a large amount of SEDP traffic from other tests
that run in parallel, starvation may occur for the topic SEDP messages
in this test. With the time-out of 30s, the test sometimes fails because
the retransmits for dropped packets are not received in time, so increasing
the timeout (that is not relevant in this test) to 180s.

Signed-off-by: Dennis Potman <dennis@zettascale.tech>