Add "Interface" attribute to network partition
The network partitions already supported unicast addresses in the
address set, allowing one to restrict the addresses to be used for
sending data to the reader to a subset of the configured interfaces.
The problem with this model is that these unicast addresses must be
those of the host, which means the configuration necessarily becomes
machine-specific.

This commit adds an "Interface" attribute that allows one to specify a
list of interface names instead.  The addresses associated with these
interfaces are then added to the set of addresses, exactly as if adding
those addresses explicitly.

Signed-off-by: Erik Boasson <eb@ilities.com>