Set visibility flag conditional and update export macro usage
The symbol visibility flag for the compiler is set depending on the
BUILD_TESTING build option. In cases tests are built, all symbols are
exported so that tests binaries can use internal functions. In case
no tests are built, only symbols for the public API and some internal
functions from DDSI and DDSRT are exported.

Some minor refactoring in idlc and ddsperf is included to limit the set
of exported functions.

Signed-off-by: Dennis Potman <dennis@zettascale.tech>