Drop unused create_topic sedp_plist where possible
This removes a vestigial bit of code for setting some topic metadata in
the discovery data.  The parameter list was ignored everywhere, and
hence had better be removed.

The only problem here is that it was exposed in the
dds_create_topic_sertype API function, and so there it remains (still as
an ignored parameter).  The documentation of the function has been
adjusted to fit reality.

Signed-off-by: Erik Boasson <eb@ilities.com>