Remove support for plists in individual writes
This commit removes code that has never been used in Cyclone DDS and was
only used for coherent set support in code inherited from the past.  The
one purpose it was used for doesn't require the generic case of
associating a full parameter list to a sample and the implementation
wasn't very clean anyway.

Signed-off-by: Erik Boasson <eb@ilities.com>