Fix assumptions that CMAKE_INSTALL_*DIR variables are relative
The CMAKE_INSTALL_*DIR variables are allowed to be absolute paths, which allows
libraries and headers to be installed into separate prefixes. This means they
cannot be naively concatenated with CMAKE_INSTALL_PREFIX. If an absolute path
is needed, CMAKE_INSTALL_FULL_*DIR should be used, while more complicated
handling is required to keep ${prefix} interpolation in pkg-config files in
the case where CMAKE_INSTALL_*DIR is relative.

Signed-off-by: Ben Wolsieffer <benwolsieffer@gmail.com>