Move cdrstream serializer
Moves the implementation of the CDR stream serializer out of DDSI
and makes it a separate component in Cyclone core. Both DDSC and
DDSI use this component for (de)serialization of application data
(via sertype_default in DDSC) and for (de)serialization of type
information for the XTypes type system (in DDSI via sertype_cdr).

Signed-off-by: Dennis Potman <dennis@zettascale.tech>