Key support in dds_cdrstream_desc_from_topic_desc
The function dds_cdrstream_desc_from_topic_desc did not copy the
keys from the topic descriptor. This is not a problem for the
current usage within cyclone c, but as this function is also
used by python (in the fuzzer tests), it needs key support.

Additionally this adds a fini function to free the allocated memory.

Signed-off-by: Dennis Potman <dennis@zettascale.tech>