Consistently use LIVELINESS also for participants
The liveliness QoS is only defined for endpoints, but Cyclone has
already been using it on participants to allow an application to
override the default lease duration for a participant.

The DDSI stack internally maintained a separate participant lease
duration, including in ddsi_plist_t.  This commit eliminates this
separate lease duration and consistently uses liveliness for encoding
all liveliness settings.

The price paid is a complication in (de)serialization of parameter
lists, because now a liveliness setting needs to be serialised as a
liveliness for SEDP but as a participant lease duration for SPDP, and
vice versa.

Signed-off-by: Erik Boasson <eb@ilities.com>