Avoid referencing dds_global in dds_whc
The built-in writers of DDSI rely on the WHC implementation in DDSC,
with the WHC implementation having no dependencies on the rest of DDSC,
except for a global mutex used to construct/destruct a freelist when the
first WHC is created or the last one deleted.  This means it is
impossible to create a DDSI participant without initialising DDSC.

Changing the mutex used to one deliberately exposed by the DDSRT layer
for process-wide initialisation purposes at least solves the practical
initialisation problem.

Signed-off-by: Erik Boasson <eb@ilities.com>