Fix CDR descr leak in type id/obj equality test
This commit replaces a call to "dds_cdrstream_desc_from_topic_desc" with
a minimal initialization of a dds_cdrstream_desc straight from the
XTypes TypeObject topic descriptor generated by IDLC, relying on a
little bit of knowledge of the type (e.g., no keys, union present) and
on a little bit of knowledge of the CDR serializer (e.g., number of
operations is not relevant for (de)serialization).

This avoids the determining the number of code words, allocating and
copying, which is rather wasteful for what really is compile-time
constant data.  The obvious alternative would have been to call
"dds_cdrstream_desc_fini".

Signed-off-by: Erik Boasson <eb@ilities.com>