IDLC option for generating CDR stream descriptor
This adds a "-f cdrstream-desc" option to the C backend, which causes it
to emit a "struct dds_cdrstream_desc" in addition to the regular topic
descriptor.  There are some small differences between the two, and the
former is usually derived from the latter at run-time.

This run-time conversion involves memory allocation and determining the
number of bytes occupied by the serializer VM instructions.  Ordinarily,
it is done only when creating a topic in the API, at which point this
little bit of overhead is not an issue.

For some internal uses, in particular concerning testing type objects
for equality, this is a bit much.  Having the required representation
available as a compile-time constant is a much more attractive option.

Signed-off-by: Erik Boasson <eb@ilities.com>