Unaligned octets-to-next-header in last submessage
Octets to next header must be a multiple of 4 in all cases *except* when
it is the last submessage in the message.  That one case got overlooked
in the message validation code.  Unfortunately, it actually gets used in
practice (at least by RTI Connext Micro).

This leads to another problem: what to do if it is not a multiple of 4,
and yet the message extends beyond it?  My reading of the specification
is that it means the message is malformed.

Signed-off-by: Erik Boasson <eb@ilities.com>