Multiple matching networks: longest netmask wins
It turns out that configurations like:

  eth0  192.168.1.1/16
  eth1  192.168.2.1/24

occur in practice, and so the logic for determining whether an address
is "local" or not will consider an address 192.168.2.2 local for both
eth0 and eth1 (I don't see how it could do anything else, it seems to be
a legal network configuration and it certainly seems to work fine in
Linux).

This commit changes the return from the first matching one (so dependent
on the order in which the interface table was filled) to the one with
the longest netmask (assuming IP networks).  That means it will now
consistently pick eth1 in the example.

Signed-off-by: Erik Boasson <eb@ilities.com>