Dynamic Type API
This introduces the implementation of the Dynamic Language binding
as defined in section 7.5.2 of the DDS XTypes specification. The API
in this implementation is roughly following the API as described in
the specification, but different choices where made at some points
(also because the approach in the spec is aimed at object oriented
languages).

This commit contains the Dynamic Type part of the dynamic language
binding, not the Dynamic Data part (which may be added later). The
type introspection part is also left out in this initial version.

Signed-off-by: Dennis Potman <dennis@zettascale.tech>