Added a introduction section and merged HelloWorld tutorials into a single tutorial
I've added a top-level section that has much the same information as the intro in the GitHub website (README.md). This also includes other high-level topics such as the data-centric architecture.

Because the HelloWorld C core DDS and C++ HelloWorld tutorials follow the same pattern, with similar explanations (except for the code examples), I've merged them into a single tutorial, with tab groups for the C and C++ specific areas.