Updates to the Core DDS (C) tutorial
The code in the Core DDS (C) tutorial did not match the subscriber.c and the publisher.c code samples in the Examples > HelloWorld. Therefore, I've replaced the code with a literalinclude so that these remain in sync. For the explanatory parts of the tutorial, I've replaced the snippets with the code from subscriber.c and the publisher.c, including line numbers so that it's easier to reference them.