Documentation edits to the Getting Started Section (#1510)
* Documentation edits to the Getting Started Section

I've made changes to the Installation and Benchmarking sections. I've also split these sections into separate files.

* Updated with review comments

* Edits to the Configuration Guide

* Edits to the DDSI Concepts page

As this was a large topic, with multiple sections, I split it into several files.

* Update docs/manual/config/ddsi-transient_behavoir.rst

Co-authored-by: Phil <25635927+poetinger@users.noreply.github.com>

* Edits to the Network configuration page

I've separated the Network configuration section into its own page.

In addition, I've updated the log on the Getting started page so that it is smaller, and has a white background.

I've also started to add index and glossary entries. The definitions for  the glossary items need to be defined - see https://github.com/orgs/ZettaScaleLabs/projects/5/views/2?visibleFields=%5B%22Title%22%2C23825276%2C%22Assignees%22%2C%22Status%22%5D

* Edits to the Data path configuration

In addition to the edits, I have put all the external html links into a separate file. Hopefully, this will make it easier to check for broken or superseded links.

* Edits to the Network partition configuration

This includes updates to cross-references, index and glossary

* Edits to the Examples pages

* Edits to the Security pages

I've split the security pages into separate files so that they appear in the left-hand TOC

* Further edits to the Configuration pages

I've separated the Network configuration into separate topics.

* Edits to the Shared memory section

I've split the Shared memory page into separate files and moved it to a Shared memory folder. I've also updated the Getting started logo to a scalable svg.

* Updated in response to review comments

Mostly, minor changes to capitalisation. Also some fixes to the Network configuration pages

* Added a introduction section and merged HelloWorld tutorials into a single tutorial

I've added a top-level section that has much the same information as the intro in the GitHub website (README.md). This also includes other high-level topics such as the data-centric architecture.

Because the HelloWorld C core DDS and C++ HelloWorld tutorials follow the same pattern, with similar explanations (except for the code examples), I've merged them into a single tutorial, with tab groups for the C and C++ specific areas.

* Updates to the Core DDS (C) tutorial

The code in the Core DDS (C) tutorial did not match the subscriber.c and the publisher.c code samples in the Examples > HelloWorld. Therefore, I've replaced the code with a literalinclude so that these remain in sync. For the explanatory parts of the tutorial, I've replaced the snippets with the code from subscriber.c and the publisher.c, including line numbers so that it's easier to reference them.

* Further updates to the C++ tutorial

I've update the C++ tutorials so that it matches the publisher.cpp and subscriber.cpp files

* Added CMake build configuration

This was in the README.md file, so I've added it to the Main documentation. Includes links to other relevant pages.

* Updated external links to OMG IDL

* Corrected a typo in the external links page

* Apply suggestions from code review

Co-authored-by: NoxPardalis <NoxPardalis@users.noreply.github.com>

* Updates to the threashold example

* Updated "Cyclone DDS" to use variable references

* Update to platforms.rst

* Update to prerequisites.rst

* Apply suggestions from code review

Co-authored-by: NoxPardalis <NoxPardalis@users.noreply.github.com>

---------

Co-authored-by: Phil <25635927+poetinger@users.noreply.github.com>
Co-authored-by: NoxPardalis <NoxPardalis@users.noreply.github.com>